+++ Assets/_Scripts/Systems/Serialization/GameData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 namespace Metroidvania.Serialization
 {
     [System.Serializable]
-    public partial class GameData
+    public partial class GameData : System.ICloneable
     {
         public int userId;
         public long lastSerialization;
@@ -12,5 +12,8 @@ namespace Metroidvania.Serialization
         {
             this.userId = userId;
         }
+
+        public GameData Clone() => this.MemberwiseClone() as GameData;
+        object System.ICloneable.Clone() => Clone();
     }
 }